+++ src/runtime/components/Modal.vue
@@ -163,14 +163,16 @@ const b24ui = computed(() => tv({ extend: tv(theme), ...(appConfig.b24ui?.modal 
         @after-leave="emits('after:leave')"
         v-on="contentEvents"
       >
-        <VisuallyHidden v-if="!!slots.content && ((title || !!slots.title) || (description || !!slots.description))">
-          <DialogTitle v-if="title || !!slots.title">
+        <VisuallyHidden v-if="(!title && !slots.title) || (!description && !slots.description) || !!slots.content">
+          <DialogTitle v-if="!title && !slots.title" />
+          <DialogTitle v-else-if="!!slots.content">
             <slot name="title">
               {{ title }}
             </slot>
           </DialogTitle>
 
-          <DialogDescription v-if="description || !!slots.description">
+          <DialogDescription v-if="!description && !slots.description" />
+          <DialogDescription v-else-if="!!slots.content">
             <slot name="description">
               {{ description }}
             </slot>
@@ -186,13 +188,15 @@ const b24ui = computed(() => tv({ extend: tv(theme), ...(appConfig.b24ui?.modal 
             <div v-if="!!slots.header || (title || !!slots.title) || (description || !!slots.description) || (props.close || !!slots.close)" data-slot="header" :class="b24ui.header({ class: uiProp?.header })">
               <slot name="header" :close="close">
                 <div data-slot="wrapper" :class="b24ui.wrapper({ class: uiProp?.wrapper })">
-                  <DialogTitle v-if="title || !!slots.title" data-slot="title" :class="b24ui.title({ class: uiProp?.title })">
+                  <DialogTitle v-if="!title && !slots.title" />
+                  <DialogTitle v-else data-slot="title" :class="b24ui.title({ class: uiProp?.title })">
                     <slot name="title">
                       {{ title }}
                     </slot>
                   </DialogTitle>
 
-                  <DialogDescription v-if="description || !!slots.description" data-slot="description" :class="b24ui.description({ class: uiProp?.description })">
+                  <DialogDescription v-if="!description && !slots.description" />
+                  <DialogDescription v-else data-slot="description" :class="b24ui.description({ class: uiProp?.description })">
                     <slot name="description">
                       {{ description }}
                     </slot>